--- src/action/index.js
@@ -1,1 +1,0 @@
-export * from './OfferAction';
