+++ src/api/get.js
@@ -1,17 +1,3 @@
 import data from './demo';
 
-class OfferAPI {
-    static async getPostList() {
-        try {
-            // const response = await fetchApi(
-            //     `whale.drkbl.com/getOffers`,
-            //     'GET'
-            // );
-            return await data; // return the dummy data for now
-        } catch (error) {
-            throw error;
-        }
-    }
-}
-
-export default OfferAPI;
+export default data;
