+++ src/components/offer/index.js
